pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fishjaw"
version = "0.1.0"
description = "Zebrafish jaw segmentation toolkit"
readme = "README.md"
requires-python = "==3.9.*"
license = {text = "MIT"}
authors = [
    {name = "Richard Lane", email = "richard.lane@bristol.ac.uk"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "monai>=1.3.0",
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scikit-image>=0.21.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "pydicom>=2.4.0",
    "tifffile>=2023.7.10",
    "trimesh>=3.22.0",
    "pyyaml>=6.0",
    "numpy-stl>=3.0.0",
    "prettytable>=3.8.0",
    "tabulate>=0.9.0",
    "torchio>=0.19.0",
    "opencv-python>=4.8.0",
    "pickle-mixin>=1.0.2",
    "jupyter>=1.0.0",
    "pyradiomics>=3.1.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "nbdime>=4.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pylint>=2.17.0",
    "black>=23.7.0",
    "mypy>=1.5.0",
    "types-PyYAML>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/JGIBristol/zebrafish_jaw_segmentation"
Repository = "https://github.com/JGIBristol/zebrafish_jaw_segmentation"

[tool.hatchling.build.targets.wheel]
packages = ["fishjaw"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["fishjaw/test/"]

[tool.pylint.messages_control]
disable = ["E0104"]
